many-to-many relationship that will never work...
room will probably one day support lists of ids. or lists, that would be
nice. but for now the alternative is to use a crossref table which
nobody seems to have any good experiences with. obviously room *can* do
M:N relationships but i do not have the time for this anymore. we have
to move on. will downgrade to the simple 1:1 notif relationship and
continue